test(web): pin the named-property clobbering fix at the JS level
Extracts the maestro-web.js Graal harness introduced with the
flt-semantics-identifier tests into a shared MaestroWebJsHarness, teaching it
to diverge a node's reflected properties from its attribute store the way
HTMLFormElement named-property access does (form.id returning a child
<input name="id"> element).

Adds WebResourceIdNamedPropertyTest on top of it: the #2944 scenario now has
a JS-level regression test instead of only the Kotlin-side coercion tests and
manual verification. Against the pre-fix maestro-web.js all four new tests
fail with the original ClassCastException failure mode.

Also strips the stray trailing whitespace this branch introduced in
CdpWebDriverTest and WebDriverTest.
