+++ maestro-client/src/test/java/maestro/drivers/FlutterWebSemanticsIdentifierTest.kt
@@ -1,17 +1,14 @@
 package maestro.drivers
 
 import com.google.common.truth.Truth.assertThat
-import org.graalvm.polyglot.Context
+import maestro.drivers.MaestroWebJsHarness.resolveResourceId
 import org.junit.jupiter.api.Test
 
 /**
- * Exercises the bundled maestro-web.js directly in a JS engine so the
- * resource-id derivation is pinned without standing up a browser.
- *
  * Flutter web puts a developer's Semantics(identifier:) on the DOM attribute
- * flt-semantics-identifier, while node.id is an internal handle of the form
- * flt-semantic-node-N that is reassigned between frames. resource-id must
- * therefore prefer the stable identifier, and must leave every existing
+ * flt-semantics-identifier, while the id attribute is an internal handle of
+ * the form flt-semantic-node-N that is reassigned between frames. resource-id
+ * must therefore prefer the stable identifier, and must leave every existing
  * (non-Flutter) resolution path untouched.
  */
 class FlutterWebSemanticsIdentifierTest {
@@ -69,125 +66,5 @@ class FlutterWebSemanticsIdentifierTest {
     fun `a node with no identifying attributes gets no resource-id`() {
         val resourceId = resolveResourceId()
         assertThat(resourceId).isNull()
-    }
-
-    // --- harness -------------------------------------------------------------
-
-    private val webScript: String by lazy {
-        requireNotNull(javaClass.classLoader.getResource("maestro-web.js")) {
-            "maestro-web.js was not found on the test classpath"
-        }.readText()
-    }
-
-    /**
-     * Builds a single element under <body>, runs maestro-web.js over it, and
-     * returns the resolved resource-id (null when the key is absent, "" when
-     * it is present but empty).
-     */
-    private fun resolveResourceId(
-        tagName: String = "flt-semantics",
-        id: String? = null,
-        ariaLabel: String? = null,
-        name: String? = null,
-        title: String? = null,
-        htmlFor: String? = null,
-        attributes: Map<String, String> = emptyMap(),
-    ): String? {
-        Context.newBuilder("js").build().use { context ->
-            context.eval("js", domScript(tagName, id, ariaLabel, name, title, htmlFor, attributes))
-            context.eval("js", webScript)
-
-            val value = context.eval(
-                "js",
-                "maestro.getContentDescription().children[0].attributes['resource-id'] ?? null",
-            )
-            return if (value.isNull) null else value.asString()
-        }
-    }
-
-    private fun domScript(
-        tagName: String,
-        id: String?,
-        ariaLabel: String?,
-        name: String?,
-        title: String?,
-        htmlFor: String?,
-        attributes: Map<String, String>,
-    ): String {
-        val directProps = buildList {
-            add("tagName: '$tagName'")
-            add("id: '${id.orEmpty()}'")
-            ariaLabel?.let { add("ariaLabel: '$it'") }
-            name?.let { add("name: '$it'") }
-            title?.let { add("title: '$it'") }
-            htmlFor?.let { add("htmlFor: '$it'") }
-        }.joinToString(",\n          ")
-
-        // id/name/title are reflected attributes: maestro-web.js reads them via
-        // Element.prototype.getAttribute, so they must live in the attribute
-        // store as well as on the node.
-        val attrEntries = buildMap {
-            id?.let { put("id", it) }
-            name?.let { put("name", it) }
-            title?.let { put("title", it) }
-            putAll(attributes)
-        }.entries.joinToString(",\n            ") { (key, v) ->
-            "'$key': { value: '$v' }"
-        }
-
-        return """
-            globalThis.Node = { TEXT_NODE: 3 };
-            globalThis.window = globalThis;
-            globalThis.innerWidth = 1024;
-            globalThis.innerHeight = 768;
-
-            // maestro-web.js reads id/name/title through Element.prototype.getAttribute
-            // (to bypass form named-property clobbering); back it with the fake
-            // attribute store.
-            globalThis.Element = {
-              prototype: {
-                getAttribute(attrName) {
-                  const attr = this.attributes[attrName];
-                  return attr === undefined ? null : attr.value;
-                },
-              },
-            };
-
-            const element = {
-              $directProps,
-              attributes: {
-                $attrEntries
-              },
-              childNodes: [{ nodeType: Node.TEXT_NODE, textContent: 'label' }],
-              children: [],
-              selected: false,
-              parentElement: null,
-              getBoundingClientRect() {
-                return { x: 0, y: 0, width: 100, height: 20 };
-              },
-            };
-
-            const body = {
-              tagName: 'body',
-              id: '',
-              attributes: {},
-              childNodes: [],
-              children: [element],
-              selected: false,
-              parentElement: null,
-              getBoundingClientRect() {
-                return { x: 0, y: 0, width: 1024, height: 768 };
-              },
-            };
-            element.parentElement = body;
-
-            globalThis.document = {
-              body: body,
-              readyState: 'complete',
-              querySelectorAll() { return []; },
-            };
-
-            globalThis.maestro = {};
-        """.trimIndent()
     }
 }
