+++ maestro-client/src/test/java/maestro/drivers/WebDriverTest.kt
@@ -147,7 +147,7 @@ class WebDriverTest {
 
         assertThat(node.attributes["resource-id"]).isInstanceOf(String::class.java)
     }
-    
+
     @Test
     fun `parseDomAsTreeNodes handles bounds as String`() {
         val dom = mapOf(
